Concat resources need to depend on target dir
With puppet-concat < 2.x, the underlying implementation used lots of
file resources.  That had the side effect that those resources would
automatically depend on the target directory.  However, with 2.x and
later, the concat resources are implemented as a native type, so we need
to add an explicit dependency on the target dir as a result.