Don't manage datadir
This change adds a new manage_datadir argument to selectively
prevent the management of the data dir.

The scenario I ran into needing this was when using a master/slave
setup. The remote data comes in via Apache with the www-data user
and it requires a setfacl entry as documented at
https://github.com/oetiker/SmokePing/wiki/FAQ#master-slave-setups
or you'll see permission denied errors in the log file.

Ideally this module would be extended to support master/slave better,
but this will help in the meantime.